Gemini_chat_with_docs/requirements.txt
streamlit>=1.28.0
google-genai>=0.1.0
python-dotenv>=1.0.0